Major Update for this Unicorn Hat LED project
This is a major step for the integration of the Unicorn Hat into Nerves.
The PWM userspace lib is derived from
[this](https://github.com/richardghirst/rpi_ws281x) implementation. Some
flavour has been added to support the 8x8 LED matrix.

There is some progress in the way the native program will be
implemented. It will be called using *Port.open()* and the communication
will be handled only through *stdin* and *stdout*. This should provide
us with an robust way for asynchronous messaging with the C integration
of the driver.

A small wrapper is given for the driver. This can be found in:
*lib/raspi_unicorn/driver*